refactor: isolate main CI tests to separate file
- Create WSConsumerMetaMainCITest.java for tests that only run in main CI
  - Call TestUtils.runInMain() in @BeforeClass to skip tests in PR CI
  - Move 7 tests that may fail in dev environments due to:
    * TDengine version differences
    * SQL syntax variations
    * Metadata format changes
- Remove 7 tests from WSConsumerMetaTest.java:
  * testAlterTable
  * testAlterMultiTableTags
  * testAlterStableTagWithFilter
  * testAlterStableTagWithRegexp
  * testCreateVirtualNormalTable
  * testCreateVirtualChildTable
  * testAlterVirtualTableAddColumnWithRef
